fix: clean the clippy backlog, and the three real defects it was hiding
`cargo clippy --all-targets --all-features -- -D warnings` passes for the
first time. Most of the 44 warnings were style, but three were reports of
code that does not do what its documentation says:

`ARCATURE_VITE_IPC` was never read. `dev_proxy::config::endpoint_from_env` is
described in five places as the startup hookup for the one-port dev topology,
and nothing called it -- the dev proxy could only be switched on by an
explicit builder call the scaffold does not make, so Vite requests would have
404'd. The builder now resolves it once at construction, and
`dev_proxy_endpoint()` is the override its doc already claimed to be.

`RedirectResponse::route()` always returns 400. It defers to a "response
mapper layer" that does not exist, so `resolve()` is dead and `.with()` drops
its flash data. The implementation is correct and the mapper is B.5.4 work, so
both methods now document the gap under a "Not yet implemented" heading rather
than failing silently at runtime.

`AppConfig` parses APP_NAME/APP_URL/APP_ENV/APP_PORT and discards all four.
Those are the inputs the security headers, the HSTS decision and release-mode
error redaction need, so the accessors are kept with a note rather than
deleted.

Also: `IntoRouteParams` appeared in the signature of a public method while
being crate-private, so no outside caller could name the bound. It is public
now, sealed by `RouteParam` having no public constructor -- which, unlike a
`Sealed` supertrait, cannot drift out of step with the impl list.

+++ src/jobs/config.rs
@@ -205,7 +205,7 @@ impl RetryPolicy {
         if self.jitter && !delay.is_zero() {
             // Full jitter: random in [0, delay]. Uses uuid::Uuid for the
             // random source (no new dependency; uuid is already a dep).
-            let cap_nanos = delay.as_nanos() as u128;
+            let cap_nanos = delay.as_nanos();
             if cap_nanos > 0 {
                 let rand = uuid::Uuid::new_v4().as_u128() % (cap_nanos + 1);
                 delay = Duration::from_nanos(u64::try_from(rand).unwrap_or(0));
